% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_split.R
\name{print.idx.split}
\alias{print.idx.split}
\title{Print method for idx.split objects}
\usage{
\method{print}{idx.split}(x, ...)
}
\arguments{
\item{x}{An object of class \code{idx.split}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original object.
}
\description{
Prints a brief summary of a train/test split result.
}
