% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_control.R
\name{print.prtree.control_cv}
\alias{print.prtree.control_cv}
\title{Print method for PRTree control objects}
\usage{
\method{print}{prtree.control_cv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{prtree.control_cv}, as returned by
\link{pr_tree_control_cv}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original object.
}
\description{
Prints a summary of the control parameters for PRTree model fitting.
}
