% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_cv_methods.R
\name{print.prtree.cv}
\alias{print.prtree.cv}
\title{Print method for PRTree Cross-Validation objects}
\usage{
\method{print}{prtree.cv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{prtree.cv}, as returned by \link{pr_tree_cv}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original object.
}
\description{
Prints a brief summary of cross-validation results for Probabilistic
Regression Trees.
}
