% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_main_methods.R
\name{print.prtree}
\alias{print.prtree}
\title{Print method for PRTree objects}
\usage{
\method{print}{prtree}(x, ...)
}
\arguments{
\item{x}{An object of class \code{prtree}, as returned by \link{pr_tree}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original object.
}
\description{
Prints a brief summary of a fitted Probabilistic Regression Tree model.
}
