% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_grid_expansion.R
\name{print.sigma_grid}
\alias{print.sigma_grid}
\title{Print method for sigma grid objects}
\usage{
\method{print}{sigma_grid}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{sigma_grid}, as returned by \link{expand_sigma_grid}.}

\item{digits}{minimal number of significant digits.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original object.
}
\description{
Prints a summary of a sigma grid generated by \link{expand_sigma_grid}.
}
