% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_split.R
\name{print.summary.idx.split}
\alias{print.summary.idx.split}
\title{Print method for summary of \code{idx.split} objects}
\usage{
\method{print}{summary.idx.split}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.idx.split}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original object.
}
\description{
Prints a detailed summary of a train/test split result.
}
