% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_cv_methods.R
\name{print.summary.prtree.cv}
\alias{print.summary.prtree.cv}
\title{Print method for summary of PRTree Cross-Validation objects}
\usage{
\method{print}{summary.prtree.cv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.prtree.cv}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original object.
}
\description{
Prints a detailed summary of cross-validation results.
}
