% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_main_methods.R
\name{print.summary.prtree}
\alias{print.summary.prtree}
\title{Print method for summary of PRTree objects}
\usage{
\method{print}{summary.prtree}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.prtree}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the original object.
}
\description{
Prints a summary of a fitted Probabilistic Regression Tree model.
}
