% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree.R
\docType{package}
\name{PRTree.Package}
\alias{PRTree-package}
\alias{PRTree.Package}
\alias{PRTree}
\title{PRTree: Probabilistic Regression Tress}
\description{
\packageDescription{PRTree}
}
\author{
Taiane Schaedler Prass \email{taianeprass@gmail.com} and Alisson Silva Neimaier \email{alissonneimaier@hotmail.com}
}
\keyword{internal}
