% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_split.R
\name{summary.idx.split}
\alias{summary.idx.split}
\title{Summary method for \code{idx.split} objects}
\usage{
\method{summary}{idx.split}(object, ...)
}
\arguments{
\item{object}{An object of class \code{idx.split}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of class \code{summary.idx.split} containing detailed statistics.
}
\description{
Provides a detailed summary of a train/test split result, including
statistics about missing values if available.
}
