% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prtree_main_methods.R
\name{summary.prtree}
\alias{summary.prtree}
\title{Summary method for PRTree objects}
\usage{
\method{summary}{prtree}(object, ...)
}
\arguments{
\item{object}{An object of class \code{prtree}, as returned by \link{pr_tree}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of class \code{summary.prtree} containing summary statistics.
}
\description{
Provides a detailed summary of a fitted Probabilistic Regression Tree model.
}
